backports: remove backport of DMA shared buffers
This DMA shared buffer backport was used for kernel < 3.5, more recent
kernel versions already have a working version of the DMA shared
buffers.
The DMA shared buffers are only used by some media drivers, but non of
these media drivers, which are actually using this API, are compiled on
kernel < 3.5. The DMA shared buffer part would need some more attention
in the next backports version, because there are some bigger changes
and instead we removed it.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>