backports: do not include net_ieee802154_lowpan in kernel >= 3.16
In kernel >= 3.16 this backport is not needed at all. When
CPTCFG_IEEE802154_6LOWPAN is not set it should not get build, we do not
have to check if we are kernel > 3.5, because in an other case
ieee802.15.4 will not get build. Kernel version 3.15 needs a different
backport than the older versions.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>