backports: fix bit_wait() and bit_wait_io()
In kernel version < 3.17 wait_on_bit() takes an action as a parameter,
which contains a schedule() or io_schedule() call in most cases. For
such calls wait_on_bit() was changed to not take an action but execute
schedule() action. This patch backports the new behavior to old kernel
versions.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>