backports: remove regulator drivers
It looks like no one is using the regulator drivers and maintaining
them costs some time because they often break. When someone needs
a specific driver please send a patch adding this one specific driver,
maintaining a small subset is not a problem, but maintaining all is a
big task.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>