backports: fix build with Ubuntu mainline kernels
We only want the UTS_UBUNTU_RELEASE_ABI var when we are on a normal
Ubuntu distribution kernel and not when we are on a Ubuntu mainline
kernel. Some of the Ubuntu mainline kernel do have an invalid octal
number in this field like 031418 and we do not want to evaluate this at
all on the Ubuntu mainline kernels.  All Ubuntu distribution kernel
have CONFIG_VERSION_SIGNATURE set so this way we can detect the which
type of kernel we are on.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>