backports: add device coredump framework
As drivers will start requiring the new framework and it has
been merged to linux/master during the 3.18 merge window, we
can include it into backports to compile if needed.

On kernels that do already have it but didn't want it built
honor that, but if building otherwise don't. This may need
to be revisited in the future.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>