backports: fold outdir as a backport identity target
The output directory is always the same as the project directory
for packaging backports but for integration these will differ,
to make introduction of integration easier to follow use the
outdir as  both project directory and target directory and fold
the existing args.outdir under the backport identity object as
both "project_dir" and "target_dir".

This commit has no functional changes.

Signed-off-by: Luis R. Rodriguez <mcgrof@suse.com>