dependencies: make IEEE802154 depend on 3.19
It looks hard to backport IEEE802154 to older kernels, because it now
uses the new member ieee802154_ptr in struct net_device. This code was
added in upstream commit 98a18b6ffc7 "netdevice: add ieee802154_ptr to
net_device"

If someone knows how to backport this stuff please do so and revert
this patch.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>