headers: fix problem when unloading mac80211 on Ubuntu 14.04
Upstream commit: 7f3a23738b01aa405c99bcf8f2be48ddf1956da4

When unloading mac80211 on the default kernel which is currently
shipped in Ubuntu 14.04 this commit causes a kernel error message. Now
we check if the list was deleted before and only do so if it was not
done before. Ubuntu probably backported this patch in their kernel.

commit 97666f936e0656a1b28c40cc33981a8db65e1a11
Author: Hauke Mehrtens <hauke@hauke-m.de>
Date:   Sun Nov 16 14:23:24 2014 +0100

    headers: delete the list in unregister_netdevice_many()

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>