backports: remove wpan-phy.h after rename to cfg802154.h
Address the renaming. cfg802154.h was already added but we forgot
to remove wpan-phy.h. This is an example type of upstream that
we should easily be able to automatically backport through
inference. Logic:

  for f in copy_list:
    if f in git_delta_renamed:
      adjust_copy_list(f)

We'd first have to build a series of types of changes from
the delta between the last linux-next tag we tracked and the
currenty evaluated one, the git_delta_renamed would contain
a list of renames that happened between the two tags.

commit 5ad60d36993596f7b3b958500f9c66c5338cd855
Author: Alexander Aring <alex.aring@gmail.com>
Date:   Sat Oct 25 09:41:02 2014 +0200

    ieee802154: move wpan-phy.h to cfg802154.h

    The wpan-phy header contains the wpan_phy struct information. Later this
    header will be have similar function like cfg80211 header. The cfg80211
    header contains the wiphy struct which is identically the wpan_phy
    struct inside 802.15.4 subsystem.

    Signed-off-by: Alexander Aring <alex.aring@gmail.com>
    Cc: Alan Ott <alan@signal11.us>
    Signed-off-by: Marcel Holtmann <marcel@holtmann.org>

Signed-off-by: Luis R. Rodriguez <mcgrof@suse.com>