backports: bump I2C_SI4713 to v3.13
This driver depends on gpio consumer.h, that depends on some
built-in kernel functionality and while we can now backport
built-in stuff we'd need to make that depend on BACKPORT_INTEGRATE
and then figure out how to backport that functionality on integration.

This doesn't seem like a good candidate to test that with though.
Let's wait for something more desirable.

Signed-off-by: Luis R. Rodriguez <mcgrof@suse.com>