backports: explicitely include string.h from device.h
string.h is usually included from the asm/ folder in most configuration,
but on some older kernels and non-x86 architecture, this inclusion is
not necessarily happening. As we're using memcpy() in the code
below, we need to be safe and make sure string.h is indeed there.

This issue has been observed on a 3.4 vendor kernels, but may be
applicable to other configurations.

Signed-off-by: Mathieu Olivari <mathieu@qca.qualcomm.com>
Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>