backports: remove patches for SKBTX_WIFI_STATUS
The tx_flags haven't been a union since 2.6.37, so since we
no longer support that kernel the '&' operation will now be
possible on all supported kernels. Use that fact to get rid
of the patch by defining SKBTX_WIFI_STATUS to 0 so the code
also disappears, but doesn't have to be patched.

While at it, clean up a double #if.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>