header: adapt new dma_buf_export() signature
In commit 817bd72 the dma_buf_export() function signature changed
from getting a long list of parameters to just getting a structure
with all the parameters in it. This patch adds a function which
translates from the new signature to the old function signature. This
function now also handles the different versions of dma_buf_export()
used in different kernel versions.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>