dependencies: deactivate DVB_RTL2830, DVB_RTL2832
The DVB_RTL2830 and DVB_RTL2832 driver are using the new regmap API
since commit b410dae and d101658 which we did not backport. Deactivate
this driver on affected kernel versions for now. In addition the
DVB_USB_RTL28XXU driver has to be deactivated because it depends on
these two modules.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>