patches: remove bt_sock_stream_recvmsg() and bt_sock_recvmsg() from ifdefs
Both bt_sock_recvmsg() and bt_sock_stream_recvmsg() are used
within bluetooth code but the backport actually required is only
for the struct proto_ops and struct proto sendmsg and recvmsg
callbacks. The bt_sock_recvmsg() and bt_sock_stream_recvmsg()
are not callbacks, they are exported but are used all within
the code we provide and carry.

This will help simplify the network/0054-struct-proto_ops-sig
collateral evolution port to Coccinelle SmPL.

1   3.0.101             [  OK  ]
2   3.1.10              [  OK  ]
3   3.2.68              [  OK  ]
4   3.3.8               [  OK  ]
5   3.4.107             [  OK  ]
6   3.5.7               [  OK  ]
7   3.6.11              [  OK  ]
8   3.7.10              [  OK  ]
9   3.8.13              [  OK  ]
10  3.9.11              [  OK  ]
11  3.10.75             [  OK  ]
12  3.11.10             [  OK  ]
13  3.12.40             [  OK  ]
14  3.13.11             [  OK  ]
15  3.14.39             [  OK  ]
16  3.15.10             [  OK  ]
17  3.16.7              [  OK  ]
18  3.17.8              [  OK  ]
19  3.18.12             [  OK  ]
20  3.19.5              [  OK  ]
21  4.0.0               [  OK  ]
22  4.1-rc1             [  OK  ]

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>