patches: change 0054-struct-proto_ops-sig backport strategy
This modifies we backport the struct proto_ops and struct proto
declarations from adding #idefs around the function prototype
arguments to just declaring our own new backport callback for
older kernels with the extra argument which calls the the
upstream function. This is done because we cannot write an
SmPL grammar patch to backport the collateral evolution as we
do it now. The backport strategy we provide now is equivalent
in terms of functionality but also a lot more clear, it also
leaves the upstream code unmodified as much as possible, this
should enable anyone reviewing the code to be able to more
easily determine exactly what is going on.

1   3.0.101             [  OK  ]
2   3.1.10              [  OK  ]
3   3.2.68              [  OK  ]
4   3.3.8               [  OK  ]
5   3.4.107             [  OK  ]
6   3.5.7               [  OK  ]
7   3.6.11              [  OK  ]
8   3.7.10              [  OK  ]
9   3.8.13              [  OK  ]
10  3.9.11              [  OK  ]
11  3.10.75             [  OK  ]
12  3.11.10             [  OK  ]
13  3.12.40             [  OK  ]
14  3.13.11             [  OK  ]
15  3.14.39             [  OK  ]
16  3.15.10             [  OK  ]
17  3.16.7              [  OK  ]
18  3.17.8              [  OK  ]
19  3.18.12             [  OK  ]
20  3.19.5              [  OK  ]
21  4.0.0               [  OK  ]
22  4.1-rc1             [  OK  ]

real    53m48.791s
user    1437m15.048s
sys     217m20.652s

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>