header: fix warning in lockdep_assert_held()
When building the kernel there are lots of warnings like this:
drivers/net/wireless/ath/ath10k/ce.c:404:21: warning: unused variable ‘ar_pci’ [-Wunused-variable]
  struct ath10k_pci *ar_pci = ath10k_pci_priv(ar);

They are caused by an empty lockdep_assert_held() which does nothing
with this variable in kernel version < 3.9. This patch replaces this
macro with the version from a recent kernel and silence this warning.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>