header: add kernel_param_{lock,unlock}() functions
In the mainline kernel the kernel module parameter locking is not
protected by a global lock any more, but each module has its own lock
now. This resulted in some changes in the api. This patch converts
calls to the new api to the old one.

This backports the changes done in this commit:
commit b51d23e4e9fea6f264d39535c2a62d1f51e7ccc3
Author: Dan Streetman <ddstreet@ieee.org>
Date:   Wed Jun 17 06:18:52 2015 +0930

    module: add per-module param_lock

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>