header: add led_trigger_remove() prototpye
The led_trigger_remove() function is exported on older kernel versions,
but there is no pubic prototype for this function. Now the first driver
makes use of this function and this patch adds the prototype.
There is no LINUX_BACKPORT() guard here because we want it to point to
the original function which is exported normally.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>