header: add compiler-gcc5.h
Upstream commit: 4760df0f39baef682278986b9f0e85ce8035cdd2

This adds support for gcc 5.

In kernel version >= 3.18 this is shipped with the kernel, in kernel
version >= 4.2 all the gcc version specific files are merged into one
file.
This patch got backported to many kernel versions, but ignore it
because it is always the same file.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>