header: add upstream version of clamp()
saa7164-encoder.c uses clamp() with constant values which fails on
kernel version < 3.18 because the code wants to creates values of the
same type and wants to assign variable values to them. The new code
works with this kind of usage, use the new code also on older kernel
versions to fix this problem.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>