header: add inet_frags_uninit_net()
inet_frags_uninit_net() was added in commit 1d6119baf06 "net: fix
percpu memory leaks" and is used by the 6lowpan part of ieee802154 in
backports.

percpu_counter_destroy() was added in kernel 3.9, so do not add this
function on older kernel versions. This is currently only used by the
6lowpan part of ieee802154 which is only activated for kernel >= 3.18.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>