backport: make media depend on kernel 3.8
The media backports are not used by so many users in my opinion and to
reduce the efforts needed to make the new changes to media introduced
by new kernel versions work on older kernel versions just make it
depend on kernel 3.8 and do not support all the older kernel versions
for media any more. This makes it possible to remove some patches and
other changes.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>