gentree: don't throw exceptions when applying patches fail
In commit 4e84c2d807d5 ("backports: move legacy and SmPL patch
application into helper"), the functionality was changed by raising
exceptions instead of return an error when applying patches fail.
This breaks git-tracker, because it expects errors to be returned in
such cases, to exit gracefully and print the error messages.

Fixes: 4e84c2d807d5 ("backports: move legacy and SmPL patch application into helper")
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>