backports: use codel and fq headers from the kernel sources
We don't need to have a local copy of codel.h anymore, so it can be
removed and copied from the kernel sources.  Additionally, it was
split into 3 different headers, so copy the other 2 as well.

Also copy the fq header files, which are used by newer wireless
subsystem versions, from the kernel sources.

Additionally, remove pkt_sched.h since it is not necessary anymore.

Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>