gentree: create Kconfig.local, use it to restrict compat
Not all the compat code is always necessary, for example code
enabled by BPAUTO_CRYPTO_SKCIPHER and by BPAUTO_RHASHTABLE is
gated on other symbols being selected. Checking against the
Kconfig symbols as it's done right now is wrong though, since
the base kernel's Kconfig symbols would be used, selecting,
for example, BPAUTO_CRYPTO_SKCIPHER when the base kernel has
CONFIG_BT set, but doing that when BT isn't even part of the
backport, or when it's disabled in the backport.

To fix this, keep track of all the local Kconfig symbols,
similar to the .local-symbols files but as a Kconfig.local,
and add for each local symbol a symbol that just mirrors the
state of the symbol as selected by the user. This effectively
allows checking the source of the symbol.

Make the relevant backported items use the new BACKPORTED_*
symbols for dependencies, instead of the original ones, thus
avoiding the above-mentioned "leakage" of kernel symbols.
When the symbol isn't even present in the backport, it'll be
simply regarded as false by the kconfig system.

Reported-by: Jouni Malinen <j@w1.fi>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>