backports: only backport kconfig.h macros for < v3.4
In kernel v4.9-rc3, config_enabled() was removed, so we can't use it
anymore.  The IS_BUILTIN macro is undefined then redefined using
config_enabled(), which doesn't work anymore.  Additionally, it's not
necessary to include all these macros if the kernel is >= v3.4.  To
fix the config_enabled() problem and to avoid backporting
unnecessarily, simply gate the backports on < v3.4.

Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>