backports: iwlwifi: access dentry name path directly for < 3.12
In 3.12, new string format types have been added (%pd and %pD), which
allows printing a dentry name or file instead of accessing them
directly.  This is really difficult (if at all feasible) to backport,
so we need to keep the old code in the iwlwifi driver for kernels that
don't have this new feature.

Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>