backports: restrict group_attr spatch properly
In cases where there are multiple instances of the ATTRIBUTE_GROUPS()
macro, or an instance thereof with a .dev_groups that doesn't actually
use it (like in devcoredump.c), the spatch wasn't correctly restricted
to make sure the two belong together, and thus made some completely
wrong modifications, linking an ATTRIBUTE_GROUPS() to a .dev_attrs it
didn't belong to at all.

Fix this by making sure that the ATTRIBUTE_GROUPS() instance and use
for .dev_groups match. This needs python indirection to create the
symbol name that ATTRIBUTE_GROUPS() creates in the C code.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>