backports: generate ifdefs for arbitrary MTU assignment
After the last patch that changed things so that the MTU
change function will only be declared when the values are
constants, this patch changes it so the ifdefs are always
added regardless.

This means that those drivers that use some expression
will now compile, with loss of MTU setting functionality.

It's necessary to put the new ifdef generation into a
separate spatch since the original one throws errors on
some drivers like airo.c.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>