backport: poll.h: add include linux/eventpoll.h
The POLL constants are getting replaced with the EPOLL constants. In
this process commit e78cd95bebd9 ("preparation to switching ->poll() to
returning EPOLL...") added an additional include to linux/poll.h. Some
drivers make use of this indirect include now and fail to compile
without it.

Older kernel versions do not define all the EPON constants so add them
if they are not available.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>