backports: speed up netdev destructor spatch by two orders of magnitude
As reported by Hauke, and analyzed by Julia, the netdev destructor
spatch is really slow because <... ...> means 0 or more hits, and
thus the spatch has to be applied on all files. However, we don't
really need it to be this way, so use <+... ...+> which allows it
to reject almost all files immediately, cutting down the processing
time by more than two orders of magnitude as it only has to touch a
handful of files, rather than all.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>