backport: fix compilation with IPV6 not set
At least in some older kernel versions (e.g. 3.10), the tcp.h header
file is not implicitly included from other headers if CONFIG_IPV6 is
not set, so we need to include it in backport-4.4.c.  Also, there is
an IPv6 structure that is used, so we should ifdef it away when IPv6
is not enabled.

Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>