backport: convert tree section names to ascii
For some reason, sometimes when the git logs contain non-ascii
characters, we get some conversion errors.  This only happens if one
of the strings we use is explicitly unicode, otherwise everything is
fine.  The only string we use that is of unicode type is the tree
names that comes from the configuration sections.  To avoid issues,
convert the strings we get from the config file into ascii before
using it.

Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>