backport: prevent unused subclass variable warning in < 3.18
In kernels earlier than 3.18, the raw_spin_lock defininition when
CONFIG_DEBUG_LOCK_ALLOC is not enabled, was dropping one of the
parameters, which caused a few "unused variable" warnings at places
where this argument was only used in the call to
raw_spin_lock_nested().  Take the latest definition which evaluates
the argument to avoid compiler warnings.

Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>