backport: make ktime_get_boottime_seconds() non-inline
With kernels < 3.17, we need to include hrtimer.h to get ktime_divns()
for the ktime_get_boottime_seconds() backport.  But we can't just
include htrtimer.h in the backport timekeeping.h because we run into
some cyclical inclusions that cause problems.  To solve that make
ktime_get_boottime_seconds() non-inline and add it to a new
backport-4.18.c file.

Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>