backports: Remove ssb and bcma
Use the versions shipped with the kernel instead.
bcma and ssb didn't change much in the last few years, most of the
changes were related to code only used on embedded devices. Removing
this reduces the number of parts we have to backport. b43, b43legacy and
brcmsmac work well with the in kernel version.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>