backports: use := assignment for subdir-ccflags-y using cc-option
In newer kernels, calling cc-option will use KBUILD_CFLAGS,
which itself gets subdir-ccflags-y added to it. Thus, we get
it referencing itself, since nothing here is evaluated until
KBUILD_CFLAGS is used: KBUILD_CFLAGS contains subdir-ccflags-y,
that in turn contains cc-option call, containing KBUILD_CFLAGS.

To avoid that, use := assignment to subdir-ccflags-y, meaning
the cc-option call is done right away and we don't end up with
KBUILD_CFLAGS referencing itself.

Note that this is the first assignment to subdir-ccflags-y, so
the += was never necessary.

This fixes build of the driver on kernels 4.15 and higher.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>