backport: add ktime_get_raw_ts64() backport for < 3.19
The getrawmonotonic64() function that is used by the
ktime_get_raw_ts64() backport was only introduced in 3.19.  To fix
compilation with earlier kernels, do the convertion from
getrawmonotonic() manually if the kernel is < 3.19.

Additionally, add timespec_to_timespec64() that we need for this
conversion (and which was only introduced in 3.17).

Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>