Revert "backports: crypto_memneq has been backported to 3.12 and 3.10"
This reverts commit 1b481348b4564970e3cfa04023f1cc22ce2b6931.

It actually broke the build to include this, as memneq.c is always
built for kernels < 3.13, and thus we get a duplicate symbol export
if we skip the backport definitions.

Easier than making the build system depend on the strange versions
is to just always use our backported version with symbol rename.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>