backports: check if genl pre_doit and post_doit are set
In the pre_doit and post_doit backports, we set the ops to our
backported version, so genetlink itself will always call it.  But then
we don't check if the actual ops are set before calling them, which
causes a call to a NULL pointer.  Fix this by checking if the ops are
actually set before calling them.

Fixes: ff6746638cf1 ("backports: genetlink: update completely")
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>