backports: Add page_frag_alloc() and __page_frag_cache_drain()
This was introduced in kernel commit 8c2dd3e4a4b ("mm: rename
__alloc_page_frag to page_frag_alloc and __free_page_frag to
page_frag_free") and commit 44fdffd7050 ("mm: add support for releasing
multiple instances of a page") and is used by the mt76 driver.
Backporting it to kernel older than 4.2 is harder so just do it till
that version.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>