header: Make napi_complete_done() return bool
In the current kernel napi_complete_done() returns a bool and this is
now used by the mt76 driver. Try to backport this functionality by
doing the check which makes the function return false already in
backports.

This was added in upstream kernel commit 364b6055738b ("net: busy-poll:
return busypolling status to drivers")

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>