kconfig: Update to KConfig to version from kernel 4.17
This updates the KConfig system used in backports to the version from
kernel 4.17. In kernel 4.18 some bigger changes to the KConfig system
were introduced which are harder to backport, so start with using this
older version.

This version now generates the zconf.lex.c and zconf.tab.c files from
the original sources, newer kernel versions will delete these files in
a make clean, after that they are now regenerated. We do not ship them
by default any more, so we need lex and yacc.

The Makefile was not copied from the mainline kernel but this is written
specifically for the backports project.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>
Cc: Anthony Wong <yp@anthonywong.net>