backports: ckmake: Limit load to number of CPUs
Instead of limiting the maximum load to a fixed number limit it to the
number of available CPUs. This should hopefully increase the speed as
less task switches are needed.

Also reduce the number of treads per build to half the number of
available CPUs. Normally multiple builds are done in parallel anyway.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>