backports: xfrm: Add skb_ext_reset()
skb_ext_reset() was added in upstream commit 174e23810cd3 ("sk_buff:
drop all skb extensions on free and skb scrubbing") and is now used by
mac80211_hwsim.

skb_ext_reset() replaced the old call to secpath_reset() in
mac80211_hwsim and other places and it also does similar things than
this function. Juts call the old secpath_reset() on older kernel
versions.

This was added to net/xfrm.h and not linux/skbuff.h because
secpath_reset() is only available in net/xfrm.h in the upstream kernel
and this is not included in linux/skbuff.h.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>