backports: significantly speed up build
Upstream commit a17ed9f1fcd12d56757bc33a275d5dfa98d7e75e

When building with /bin/sh -> bash, things are SUPER slow (at least
for me), because bash takes a LOOONG time to look at the environment
variables (and we typically have around 6k), adding over a second to
each bash invocation for me.

The reason we export them is that we need them in all of the sub-
makes, and those only read auto.conf, which we can't change.

Work around this by overriding 'make' itself, and using --eval to
read *our* .config file into each make that gets called. This way,
the variables are present in all make invocations in the same way
as they would be through the environment, but don't get passed to
shell invocations.

If --eval is not supported, keep doing what we did before.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>