backports: Remove coredump subsystem backport
This is not needed on kernel 4.0 and later. These kernel versions
already ship a compatible coredump system.

Keep the BPAUTO_WANT_DEV_COREDUMP. Drivers can use WANT_DEV_COREDUMP if
they want to make it possible to select DEV_COREDUMP, but this is not
working with backports. This leaves the dummy option.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>