headers: Add rfkill_set_hw_state_reason()
This adds the new function rfkill_set_hw_state_reason() from kernel 5.11
and maps it to the old rfkill_set_hw_state() by dropping the reason.
kernel 5.11 also moved some definitions to the uapi/linux/rfkill.h, but
on kernel versions < 5.11 it is not sufficient to only include the uapi
version.
This is used by the core wireless system.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>